+++ config/settings.php
@@ -82,6 +82,12 @@ return [
                     'id' => 'setting_pdc_use_feedback_form',
                     'type' => 'checkbox',
                 ],
+				'pdc_use_theme_tiles' => [
+                    'name' => __('Theme tiles', 'pdc-base'),
+                    'desc' => __('Enable theme tiles.', 'pdc-base'),
+                    'id' => 'setting_pdc_use_theme_tiles',
+                    'type' => 'checkbox',
+                ],
                 'upl_heading' => [
                     'type' => 'heading',
                     'name' => __('The Uniform Product list of names (upl)', 'pdc-base'),